applied aggregate pattern and encapsulation
