+++ COVID19Py/covid19.py
@@ -1,127 +1,54 @@
 from typing import Dict, List
-import requests
-import json
+from .fetchdata import FetchData
+
 
 class COVID19(object):
-    default_url = "https://covid-tracker-us.herokuapp.com"
-    url = ""
-    data_source = ""
-    previousData = None
-    latestData = None
-    _valid_data_sources = []
-
-    mirrors_source = "https://raw.github.com/Kamaropoulos/COVID19Py/master/mirrors.json"
-    mirrors = None
-
     def __init__(self, url="https://covid-tracker-us.herokuapp.com", data_source='jhu'):
-        # Skip mirror checking if custom url was passed
-        if url == self.default_url:
-            # Load mirrors
-            response = requests.get(self.mirrors_source)
-            response.raise_for_status()
-            self.mirrors = response.json()
+        """
+        Initialize the root entity
 
-            # Try to get sources as a test
-            for mirror in self.mirrors:
-                # Set URL of mirror
-                self.url = mirror["url"]
-                result = None
-                try:
-                    result = self._getSources()
-                except Exception as e:
-                    # URL did not work, reset it and move on
-                    self.url = ""
-                    continue
+        :param url: str: url of the source to retrieve data
+        :param data_source: str: name of the source to retrieve data from
+        """
+        try:
+            self.rootIdentity = url
+            self.dataSource = FetchData(self.rootIdentity, data_source)
 
-                # TODO: Should have a better health-check, this is way too hacky...
-                if "jhu" in result:
-                    # We found a mirror that worked just fine, let's stick with it
-                    break
-
-                # None of the mirrors worked. Raise an error to inform the user.
-                raise RuntimeError("No available API mirror was found.")
-
-        else:
-            self.url = url
-
-        self._valid_data_sources = self._getSources()
-        if data_source not in self._valid_data_sources:
-            raise ValueError("Invalid data source. Expected one of: %s" % self._valid_data_sources)
-        self.data_source = data_source
-
-    def _update(self, timelines):
-        latest = self.getLatest()
-        locations = self.getLocations(timelines)
-        if self.latestData:
-            self.previousData = self.latestData
-        self.latestData = {
-            "latest": latest,
-            "locations": locations
-        }
-
-    def _getSources(self):
-        response = requests.get(self.url + "/v2/sources")
-        response.raise_for_status()
-        return response.json()["sources"]
-
-    def _request(self, endpoint, params=None):
-        if params is None:
-            params = {}
-        response = requests.get(self.url + endpoint, {**params, "source":self.data_source})
-        response.raise_for_status()
-        return response.json()
+        except:
+            raise Exception("Error connecting to source!")
 
     def getAll(self, timelines=False):
-        self._update(timelines)
-        return self.latestData
+        """
+        This methods returns the all the latest data, optionally with the timeline too
+
+        :param timelines: bool: True or False value to receive timeline of the data
+        :return: Dict[str: int]: containing all the available data with one command.
+
+        """
+        return self.dataSource.getAll(timelines=timelines)
 
     def getLatestChanges(self):
-        changes = None
-        if self.previousData:
-            changes = {
-                "confirmed": self.latestData["latest"]["confirmed"] - self.latestData["latest"]["confirmed"],
-                "deaths": self.latestData["latest"]["deaths"] - self.latestData["latest"]["deaths"],
-                "recovered": self.latestData["latest"]["recovered"] - self.latestData["latest"]["recovered"],
-            }
-        else:
-            changes = {
-                "confirmed": 0,
-                "deaths": 0,
-                "recovered": 0,
-            }
-        return changes
+        """
+        :return: Dict[str: int]: returns the latest changes in number of confirmed, deaths, and recovered cases
+                                 since the last retrieval of data.
+        """
+        return self.dataSource.getLatestChanges()
 
     def getLatest(self) -> List[Dict[str, int]]:
         """
         :return: The latest amount of total confirmed cases, deaths, and recoveries.
         """
-        data = self._request("/v2/latest")
-        return data["latest"]
+        return self.dataSource.getLatest()
 
     def getLocations(self, timelines=False, rank_by: str = None) -> List[Dict]:
         """
         Gets all locations affected by COVID-19, as well as latest case data.
+
         :param timelines: Whether timeline information should be returned as well.
         :param rank_by: Category to rank results by. ex: confirmed
         :return: List of dictionaries representing all affected locations.
         """
-        data = None
-        if timelines:
-            data = self._request("/v2/locations", {"timelines": str(timelines).lower()})
-        else:
-            data = self._request("/v2/locations")
-
-        data = data["locations"]
-        
-        ranking_criteria = ['confirmed', 'deaths', 'recovered']
-        if rank_by is not None:
-            if rank_by not in ranking_criteria:
-                raise ValueError("Invalid ranking criteria. Expected one of: %s" % ranking_criteria)
-
-            ranked = sorted(data, key=lambda i: i['latest'][rank_by], reverse=True)
-            data = ranked
-
-        return data
+        return self.dataSource.getLocations(timelines=timelines, rank_by=rank_by)
 
     def getLocationByCountryCode(self, country_code, timelines=False) -> List[Dict]:
         """
@@ -129,30 +56,19 @@ class COVID19(object):
         :param timelines: Whether timeline information should be returned as well.
         :return: A list of areas that correspond to the country_code. If the country_code is invalid, it returns an empty list.
         """
-        data = None
-        if timelines:
-            data = self._request("/v2/locations", {"country_code": country_code, "timelines": str(timelines).lower()})
-        else:
-            data = self._request("/v2/locations", {"country_code": country_code})
-        return data["locations"]
-    
+        return self.dataSource.getLocationByCountryCode(country_code=country_code, timelines=timelines)
+
     def getLocationByCountry(self, country, timelines=False) -> List[Dict]:
         """
         :param country: String denoting name of the country
         :param timelines: Whether timeline information should be returned as well.
         :return: A list of areas that correspond to the country name. If the country is invalid, it returns an empty list.
         """
-        data = None
-        if timelines:
-            data = self._request("/v2/locations", {"country": country, "timelines": str(timelines).lower()})
-        else:
-            data = self._request("/v2/locations", {"country": country})
-        return data["locations"]
+        return self.dataSource.getLocationByCountry(country=country, timelines=timelines)
 
     def getLocationById(self, country_id: int):
         """
         :param country_id: Country Id, an int
         :return: A dictionary with case information for the specified location.
         """
-        data = self._request("/v2/locations/" + str(country_id))
-        return data["location"]
+        return self.dataSource.getLocationById(country_id=country_id)
